Fixes several wiring bugs

+++ stack/core/src/main/java/org/apache/usergrid/corepersistence/index/IndexService.java
@@ -41,9 +41,11 @@ public interface IndexService {
      * @param applicationScope The scope of the entity
      * @param entity The entity
      *
-     * @return An observable with the count of every
+     * @return An observable with the count of every batch executed to index the entity.  Note that this a cold observable
+     * and must be subscribed to in order to perform the operation.  This also makes no assumptions on scheduling.  It is up to the caller
+     * to assign the correct scheduler to the observable
      */
-    Observable<Integer> indexEntity( final ApplicationScope applicationScope, final Entity entity );
+    Observable<Long> indexEntity( final ApplicationScope applicationScope, final Entity entity );
